pyproject.toml
[project]
name = "psychologist-aggregator"
version = "0.1.0"
description = ""
authors = [
    {name = "Maksim Lakovich",email = "maks.lakovich@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "django (>=5.2.7,<6.0.0)",
    "psycopg2-binary (>=2.9.11,<3.0.0)",
    "pillow (>=12.0.0,<13.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "python-slugify (>=8.0.4,<9.0.0)",
    "django-phonenumber-field (>=8.3.0,<9.0.0)",
    "phonenumbers (>=9.0.17,<10.0.0)",
    "django-timezone-field (>=7.1,<8.0)",
    "django-countries (>=7.6.1,<8.0.0)",
    "djangorestframework-simplejwt (>=5.5.1,<6.0.0)",
    "django-filter (>=25.2,<26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.3.0"
mypy = "^1.18.2"
isort = "^7.0.0"
black = "^25.9.0"

